+++ src/transports/encodePolyfill.ts
@@ -1,14 +1,20 @@
-import { CAP_GLOBAL_OBJ } from '../utils/webViewUrl';
+import { getMainCarrier, SDK_VERSION } from '@sentry/core';
+import type { CAP_GLOBAL_OBJ } from 'src/utils/webViewUrl';
+
 import { utf8ToBytes } from '../vendor';
 
 export const useEncodePolyfill = (): void => {
-  if (!CAP_GLOBAL_OBJ.__SENTRY__) {
-    (CAP_GLOBAL_OBJ.__SENTRY__ as Partial<
+  const globalCarriers = getMainCarrier().__SENTRY__;
+  if (!globalCarriers) {
+    (globalCarriers as Partial<
       (typeof CAP_GLOBAL_OBJ)['__SENTRY__']
     >) = {};
   }
+  const capacitorSiblingCarrier = globalCarriers?.[SDK_VERSION];
 
-  CAP_GLOBAL_OBJ.__SENTRY__.encodePolyfill = encodePolyfill;
+  if (capacitorSiblingCarrier) {
+    capacitorSiblingCarrier.encodePolyfill = encodePolyfill;
+  }
 };
 
 export const encodePolyfill = (text: string): Uint8Array => {