+++ CHANGELOG.md
@@ -2,13 +2,18 @@
 
 <!-- prettier-ignore-start -->
 > [!IMPORTANT]
-> If you are upgrading to the `1.x` versions of the SDK from `0.x`, make sure you follow our
+> If you are upgrading to the `2.x` versions of the SDK from `1.x` or lower, make sure you follow our
 > [migration guide](https://docs.sentry.io/platforms/javascript/guides/capacitor/migration/) first.
 <!-- prettier-ignore-end -->
 
 ## 1.4.0
 
 ### Dependencies
+
+- Bump JavaScript SDKs from v8.55.0 to v9.0.0 ([#840](https://github.com/getsentry/sentry-capacitor/pull/840))
+
+  - [changelog](https://github.com/getsentry/sentry-javascript/blob/9.0.0/CHANGELOG.md#900)
+  - [diff](https://github.com/getsentry/sentry-javascript/compare/8.55.0...9.0.0)
 
 - Bump Sentry Javascript to 8.55.0 ([#870](https://github.com/getsentry/sentry-capacitor/pull/870))
   - [changelog](https://github.com/getsentry/sentry-javascript/blob/8.55.0/CHANGELOG.md)