+++ CHANGELOG.md
@@ -10,10 +10,14 @@
 
 ### Dependencies
 
-- Bump JavaScript SDKs from v8.42.1 to v9.0.0 ([#840](https://github.com/getsentry/sentry-capacitor/pull/840))
+- Bump JavaScript SDKs from v8.55.0 to v9.0.0 ([#840](https://github.com/getsentry/sentry-capacitor/pull/840))
 
   - [changelog](https://github.com/getsentry/sentry-javascript/blob/9.0.0/CHANGELOG.md#900)
-  - [diff](https://github.com/getsentry/sentry-javascript/compare/8.42.0...9.0.0)
+  - [diff](https://github.com/getsentry/sentry-javascript/compare/8.55.0...9.0.0)
+
+- Bump Sentry Javascript to 8.55.0 ([#870](https://github.com/getsentry/sentry-capacitor/pull/870))
+  - [changelog](https://github.com/getsentry/sentry-javascript/blob/8.55.0/CHANGELOG.md)
+  - [diff](https://github.com/getsentry/sentry-javascript/compare/8.42.0...8.55.0)
 
 ## 1.3.0